// Web/ClientApp/services/generated/species-info-model.ts
//----------------------
// <auto-generated>
//     Generated by Ozytis Model Generator  DO NOT EDIT!
// </auto-generated>
//----------------------

import { SpeciesModel } from "./species-model";
export interface SpeciesInfoModel extends SpeciesModel
{
    pictures: string[];
    habitat: string;
    usage: string;
    description: string;
}